Moved components with css into a folder
